Split metadata tag completion logic
